FIx solution for CF_1A and CF_230A, also rename the files

+++ src/codeforces/CF_1A_Theatre_Square.kt
@@ -2,18 +2,14 @@ package codeforces
 
 import java.util.*
 
-/**
- * Created by rohmanhakim on 3/2/18.
- */
-
 // http://codeforces.com/problemset/problem/1/A
-fun leastFlagStones(squareLength: Int, paveLength: Int): Int =
+fun leastFlagStones(squareLength: Long, paveLength: Long): Long =
         if((squareLength % paveLength) > 0) (squareLength / paveLength) + 1 else (squareLength / paveLength)
 
 fun main(args: Array<String>) = with(Scanner(System.`in`)) {
-    val n = nextInt()
-    val m = nextInt()
-    val a = nextInt()
+    val n = nextLong()
+    val m = nextLong()
+    val a = nextLong()
 
     println(leastFlagStones(n,a) * leastFlagStones(m,a))
 }
